fix(desktop): use direct imports for split infra modules
