+++ apps/desktop/src/hooks/access/mcp/connectors/use-connectors.ts
@@ -1,19 +1,19 @@
 import type { QueryClient } from "@tanstack/react-query";
 import { useQuery } from "@tanstack/react-query";
-import { loadConnectorPaneData } from "@/lib/workflows/mcp/connector-persistence";
+import { loadCloudConnectorPaneData } from "@/lib/workflows/mcp/connector-catalog-persistence";
 import { mcpConnectorsKey } from "./query-keys";
 
 export function useConnectors() {
   return useQuery({
     queryKey: mcpConnectorsKey(),
-    queryFn: loadConnectorPaneData,
+    queryFn: loadCloudConnectorPaneData,
   });
 }
 
 export async function refreshMcpConnectorsQuery(queryClient: QueryClient) {
   return queryClient.fetchQuery({
     queryKey: mcpConnectorsKey(),
-    queryFn: loadConnectorPaneData,
+    queryFn: loadCloudConnectorPaneData,
     staleTime: 0,
   });
 }
