--- apps/desktop/src/lib/integrations/auth/orchestration.ts
@@ -1,8 +1,0 @@
-export { bootstrapAuth } from "./orchestration-bootstrap";
-export { handleDesktopCallbackUrl } from "./orchestration-callback";
-export {
-  linkDesktopProvider,
-  signInWithGitHub,
-  signOut,
-} from "./orchestration-provider-flow";
-export type { AuthOrchestrationDeps } from "./orchestration-effects";
